+++ docs/CHANGELOG.md
@@ -7,6 +7,11 @@ title: Changelog
 
 ## main
 
+* Add `--stimulus` flag to the component generator. Generates a Stimulus controller alongside the component.
+* Add config option `config.view_component.generate_stimulus_controller` to always generate a Stimulus controller.
+
+    *Sebastien Auriault*
+
 ## 2.37.0
 
 * Clarify slots example in docs to reduce naming confusion.